add Ubuntu 19.10 Eoan Ermine (who came up with that unspellable name)

+++ build_all.sh
@@ -2,11 +2,14 @@
 
 OUT=`pwd`/out
 
+if [ -z "$1" ]; then
+	echo "Example usage: build_all.sh buster debian"
+	exit 1
+fi
+
 mkdir -p ${OUT}
-for DEB in stretch; do
-	[ ! -z "${OUT}" -a -x ${OUT}/${DEB} ] && rm -rf ${OUT}/${DEB}
-	mkdir -p ${OUT}/${DEB}
-	for PKG in proj6 libgeotiff5 ogdi4 spatialite gdal3; do
-		bash build.sh $PKG $DEB
-	done
+[ ! -z "${OUT}" -a -x ${OUT}/${DEB} ] && rm -rf ${OUT}/${DEB}
+mkdir -p ${OUT}/${DEB}
+for PKG in proj6 libgeotiff5 ogdi4 spatialite gdal3; do
+	bash build.sh $PKG $1 $2
 done
